Moving HslColor from Particles to Core
