+++ Source/MonoGame.Extended.Particles/Modifiers/Interpolators/ColorInterpolator.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using MonoGame.Extended;
 
 namespace MonoGame.Extended.Particles.Modifiers.Interpolators
 {
