+++ Source/MonoGame.Extended.Particles/Modifiers/VelocityColorModifier.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using MonoGame.Extended;
 
 namespace MonoGame.Extended.Particles.Modifiers
 {
