+++ Source/MonoGame.Extended.Particles/Particle.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 using Microsoft.Xna.Framework;
+using MonoGame.Extended;
 
 namespace MonoGame.Extended.Particles
 {
