+++ Source/MonoGame.Extended.Particles/ParticleReleaseParameters.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
+using MonoGame.Extended;
 
 namespace MonoGame.Extended.Particles
 {
