+++ Source/MonoGame.Extended/Serialization/HslColorJsonConverter.cs
